refactor/feat: change logic database and create reading route
