+++ backend/src/__tests__/books.test.ts
@@ -1,4 +1,4 @@
-import { getAllBooks, getBook } from '../services/books';
+import { getAllBooks, getBook } from '../services/book';
 import {
   mockSearchResponse,
   mockBookDetails,
