+++ backend/src/config/database.ts
@@ -1,0 +1,4 @@
+import { createClient } from '@supabase/supabase-js';
+import { supabaseUrl, supabaseKey } from '../config/env';
+
+export const supabase = createClient(supabaseUrl, supabaseKey);
