+++ backend/src/controllers/book/index.ts
@@ -1,1 +1,1 @@
-export * from './auth-controller';
+export * from './book-controller';
