--- backend/src/controllers/wishlist/index.ts
@@ -1,1 +1,0 @@
-export * from './wishlist-controller';
