--- backend/src/routes/books.ts
@@ -1,9 +1,0 @@
-import { Router } from 'express';
-import { getAllBooksController, getBookControler } from '../controllers/book';
-
-const router = Router();
-
-router.get('/', getAllBooksController);
-router.get('/:id', getBookControler);
-
-export default router;
