--- backend/src/services/books/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * from './books-service';
